Fixed p15.cpp integer overflow

+++ Solutions/p15.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <vector>
+#include <cinttypes>
 
 // 20/10/2018
 // This one took a while to figure out how to solve
@@ -7,9 +8,9 @@
 class PascalTriangle
 {
 private:
-    std::vector<unsigned> b;
-    const unsigned int width = 0;
-    const unsigned int height = 0;
+    std::vector<uint64_t> b;
+    const uint32_t width = 0;
+    const uint32_t height = 0;
 public:
     PascalTriangle(unsigned w, unsigned h)
         : width(w + 1), height(h + 1)
@@ -24,8 +25,8 @@ public:
             }
         }
     }
-    std::vector<unsigned>& getTriangle() { return this->b; }
-    unsigned& getValueAtCoords(unsigned x, unsigned y) { return this->b.at(y * this->width + x); }
+    std::vector<uint64_t>& getTriangle() { return this->b; }
+    uint64_t& getValueAtCoords(unsigned x, unsigned y) { return this->b.at(y * this->width + x); }
 };
 
 int main()
